Replace utils.py
@AkintolaOlasubomi  had uploaded the file outside of the required directory. (project-analysis/analysis). I replaced the previous file by moving the new one into the directory.
